requirements.txt
# File location: requirements.txt

# Core ML/DL
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
pytorch-lightning>=2.1.0
torchmetrics>=1.2.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Export/Deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional utilities
tqdm>=4.65.0
rich>=13.5.0
typer>=0.9.0